Remove OwnerId from certain layouts
I'm not entirely sure why this wasn't a problem originally, but now when
I attempt to do `sfdx force:source:push` to a fresh scratch org it gives
errors about the OwnerId field not existing on the following objects:
* Class
* Deferral
* Sponsorship

My best guess is that when I originally created them with Schema Builder
/ Object Manager in the scratch org and then downloaded their XML using
`sfdx force:source:pull`, that they had an OwnerId and so the associated
layout was created with that field.

Later, when these objects were turned into children in a master-detail
relationship they lost the OwnerId field. And it's only now, when
pushing to an empty scratch org, that they Layout definitions for each
of the custom objects were actually being pushed (since they didn't
change after the initial pull).

My solution to that issue was to remove the OwnerId field from the
Layout entirely.